Added more Arquillian Graphene tests. Progressing on adding Graphene tests for all JSF/UI aspects.
The spectacle of astrology in the White House -- the governing center of
the world's greatest scientific and military power -- is so appalling
that
it defies understanding and provides grounds for great fright.  The
easiest
response is to laugh it off, and to indulge in wisecracks about Civil
Service ratings for horoscope makers and palm readers and whether Reagan
asked Mikhail Gorbachev for his sign.  A contagious good cheer is the
hallmark of this presidency, even when the most dismal matters are
concerned.
But this time, it isn't funny.  It's plain scary.
-- Daniel S. Greenberg, Editor, _Science and Government Report_, writing
in
   "Newsday", May 5, 1988